@@ -7,21 +7,20 @@
     "build": "vue-cli-service build"
   },
   "dependencies": {
-    "autoprefixer": "^9.6.0",
-    "es6-promise": "^4.2.6",
-    "trianglify": "^2.0.0",
-    "vue": "^2.6.6",
-    "vue-lazyload": "^1.2.6",
-    "vue-router": "^3.0.1",
-    "vuex": "^3.1.0"
+    "autoprefixer": "^10.4.12",
+    "es6-promise": "^4.2.8",
+    "vue": "^3.2.41",
+    "vue-router": "^4.1.6",
+    "vue3-lazyload": "^0.3.6",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.4.0",
-    "@vue/cli-service": "^3.4.0",
-    "cname-webpack-plugin": "^1.0.5",
-    "node-sass": "^4.12.0",
-    "sass-loader": "^7.1.0",
-    "vue-template-compiler": "^2.5.21",
-    "vueperslides": "^1.18.0"
+    "@vue/cli-plugin-babel": "^5.0.8",
+    "@vue/cli-service": "^5.0.8",
+    "cname-webpack-plugin": "^3.0.0",
+    "node-sass": "^7.0.3",
+    "sass-loader": "^13.1.0",
+    "vue-template-compiler": "^2.7.13",
+    "vueperslides": "^3.5.1"
   }
 }
