@@ -1,10 +1,7 @@
-import Vue from 'vue'
-import Router from 'vue-router'
+import {createRouter, createWebHistory} from 'vue-router'
 import Home from './views/Home.vue'
 
-Vue.use(Router)
-
-export default new Router({
+const createAppRouter = () => createRouter({
   routes: [
     {
       path: '/',
@@ -15,14 +12,12 @@ export default new Router({
       path: '/project',
       name: 'project',
       component: () => import('./views/Project.vue')
-    },
-    {
-      path: '/about',
-      name: 'about',
-      component: () => import('./views/About.vue')
     }
   ],
+  history: createWebHistory(),
   scrollBehavior () {
     return { x: 0, y: 0 }
   }
 })
+
+export default createAppRouter
