@@ -1,94 +1,67 @@
 <template>
-  <div class="projects-list">
-    <div
-      class="projects-list__project"
-      v-for="(image, i) in nestedImages"
-      :key="i" :class="{ 'no-hover' : image.link == undefined }">
-      
-      <router-link :to="image.link" v-if="image.link">
-        <div
-          v-lazy:background-image="image.isLocalImg 
-            ? getImgUrl(image.URL)
-            : image.URL"
-          class="projects-list__project-bg"
-          lazy="loading"
-          @click="handleClick(image)">
-        </div>
+  <div class="projects-container">
+    <div class="projects-list">
+      <div
+        class="projects-list__project"
+        v-for="(image, i) in images"
+        :key="i">
+        
+        <router-link class="projects-list__project-image" :to="image.link">
+          <div
+            v-lazy:background-image="image.isLocalImg 
+              ? getImgUrl(image.URL)
+              : image.URL"
+            class="projects-list__project-bg"
+            lazy="loading"
+            @click="handleClick(image)">
+          </div>
 
-        <p @click="handleClick(image)"
-          class="projects-list__project-bg--text">
-          {{ image.section }}
-        </p>
-      </router-link>
+          <p @click="handleClick(image)"
+            class="projects-list__project-bg--text">
+            {{ image.section }}
+          </p>
+        </router-link>
 
-      <canvas :class="`canvas-${image.id}`" v-if="!image.link"></canvas>
-
+      </div>
     </div>
   </div>
 </template>
 <script>
-import { mapActions } from 'vuex';
-import { patternGen } from '../helpers';
+import { mapActions } from 'vuex'
 
 export default {
   data() {
     return {
-      nestedImages: [],
+      images: [],
     }
   },
 	props: {
-		images: Array
+		projects: Array
   },
   created() {
-    this.nestedImages = this.getShuffledImages()
-  },
-  mounted() {
-    this.updateCanvasPatterns()
+    this.images = this.getProjectImages()
   },
   methods: {
     ...mapActions([
 			'updateProjectPageData'
 		]),
     getImgUrl(url) {
-			let images = require.context('../assets/images', true, /\.jpg|\.png$/)
-			return images(url)
+			let imagesDirContext = require.context('../assets/images', true, /\.jpg|\.png$/)
+			return imagesDirContext(url)
     },
     handleClick(image) {
       this.updateProjectPageData(image)
     },
-    getShuffledImages() {
-      let imgs = []
-      let patternNum = 0;
+    getProjectImages() {
+      let projectImages = []
 
-      for (let i = 0; i < this.images.length; i++) {
-        if (i > 0) {
-          for(let j = 0; j < 2; j++) {
-            if(this.images[i].images[j]) {
-              imgs.push(this.images[i].images[j])
-            }
-          }
-        }
+      for (let i = 0; i < this.projects.length; i++) {
+        const project = this.projects[i]
+        projectImages.push(project.images[0])
       }
 
-      imgs.unshift(...this.images[0].images)
-      
-      for(let i = 0; i < imgs.length; i++) {
-        if(i%2!==0) {
-          imgs.splice(i, 0, {id: patternNum})
-          patternNum++
-        }
-      }
-
-      imgs.push({id: patternNum})
-
-      return imgs
+      return projectImages
     },
-    updateCanvasPatterns() {
-      for (let i = 0; i < 9; i++) {
-        const pattern = patternGen()
-        pattern.canvas(document.querySelector(`.canvas-${i}`))
-      }
-    }
   }
 }
 </script>
@@ -96,40 +69,34 @@ export default {
 @import "../lib/_colors.scss";
 @import "../lib/_mixins.scss";
 
+.projects-container {
+
+}
+
 .projects-list {
   display: flex;
   flex-wrap: wrap;
   margin: 0;
   padding: 0;
-  height: 100%;
-  overflow: scroll;
-  scroll-snap-type: both mandatory;
-  grid-gap: 10px;
   scroll-padding: 1rem;
 
   @include tablet {
-    display: grid;
-    padding: 10px;
-    grid-template-columns: repeat(3, 33%);
-    grid-template-rows: repeat(12, 20vh);
   }
 
   &__project {
     width: 100%;
-    height: 100px;
-    margin-bottom: 10px;
-    list-style: none;
-    scroll-snap-align: start;
     position: relative;
     transition: 400ms;
     background-color: #f9853325;
     overflow: hidden;
+    border-radius: 5px;
+    margin-bottom: 10px;
 
     @include tablet {
       background-color: #f98433;
-      margin-bottom: 0;
-      height: auto;
-      width: auto;
+      width: calc(50% - 20px);
+      height: 500px;
+      margin: 10px;
     }
 
     &:hover {
@@ -152,37 +119,29 @@ export default {
       }
     }
 
-    &.no-hover {
-      opacity: 0.8;
-
-      .projects-list__project-bg {
-        opacity: 1;
-        transform: scale(1)
-      }
-
-      .projects-list__project-bg--text {
-        opacity: 0;
-      }
-    }
-
     & a {
       text-decoration: none;
       color: white;
     }
+  }
 
-    canvas {
-      height: 100%;
-      width: 100%;
-    }
+  &__project-image {
+    height: 100px;
+    width: 200%;
   }
 
   &__project-bg {
     transition: 300ms;
-    height: 100%;
-    width: 100%;
+    height: 230px;
     background-size: cover;
     background-repeat: no-repeat;
     background-position: center;
+    opacity: 0.4;
+
+    @include tablet {
+      opacity: 1;
+      height: 500px;
+    }
 
     &[lazy=loading] {
       background-color: $orange;
@@ -198,7 +157,12 @@ export default {
       transform: translate(-50%, -50%);
       margin: 0;
       color: #fff;
-      opacity: 0;
+      opacity: 1;
+      text-shadow: 0px 1px 40px #575757;
+
+      @include tablet {
+        opacity: 0;
+      }
     }
   }
 
