@@ -3,9 +3,9 @@ import Trianglify from 'trianglify'
 
 export const patternGen = () => {
 	return Trianglify({
-		height: window.innerHeight,
-		width: window.innerWidth,
-		cell_size: 800 + (Math.random() * 10),
-		x_colors: colors
+		height: 100,
+		width: 2000,
+		cellSize: 200 + (Math.random() * 10),
+		xColors: colors
 	})
 }
